Scrape US Census Data

Takes in a city/state and outputs demographic data from the Missouri Census Data Center

getCensus(cities, radius)

Arguments

cities
list of cities to be passed

radius
numeric number used to create a boundary around the city coordinates

Value

Data frame containing demographic data

Examples

radius <- 2
cities <- c("Chicago", "New York")
df <- getCensus(cities, radius)

#> Information from URL : http://maps.googleapis.com/maps/api/geocode/json?address=Chicago&sensor=false

#> Information from URL : http://maps.googleapis.com/maps/api/geocode/json?address=New%20York&sensor=false

#> [[1]]
#> [1] "http://mcdc.missouri.edu/cgi-bin/broker?_PROGRAM=websas.caps10acsb.sas&SERVICE=appdev&sitename=Chicago&longitude=-87.6297982&latitude=41.8781136&radii=2&dprofile=on&eprofile=on&sprofile=on&hprofile=on&cntypops=on&printdetail=on&_debug="
#> 
#> [[2]]
#> [1] "http://mcdc.missouri.edu/cgi-bin/broker?_PROGRAM=websas.caps10acsb.sas&SERVICE=appdev&sitename=NewYork&longitude=-74.0059728&latitude=40.7127753&radii=2&dprofile=on&eprofile=on&sprofile=on&hprofile=on&cntypops=on&printdetail=on&_debug="
#> 

head(df)

#>   sitename radius   Longitude   Latitude TotPopSF1 AreaSQMI LandSQMI  IntPtLat
#> 1  Chicago      2 -87.6297982 41.8781136   172,300 10.67397     9.12 41.882238
#>     IntPtLon vintage period  TotPop Age0_4 pctAge0_4 Age5_9 pctAge5_9 Age10_14
#> 1 -87.634391    2015    5yr 188,040  8,650       4.6  4,172       2.2    3,076
#>   pctAge10_14 Age15_19 pctAge15_19 Age20_24 pctAge20_24 Age25_34 pctAge25_34
#> 1         1.6    8,062         4.3   18,371         9.8   61,450        32.7
#>   Age35_44 pctAge35_44 Age45_54 pctAge45_54 Age55_59 pctAge55_59 Age60_64
#> 1   27,147        14.4   18,517         9.9    9,374         5.0    9,322
#>   pctAge60_64 Age65_74 pctAge65_74 Age75_84 pctAge75_84 Over85 pctOver85
#> 1         5.0   12,212         6.5    5,917         3.1  1,770       0.9
#>   MedianAge  Over15 pctOver15 Under18 pctUnder18  Over18 pctOver18  Over21
#> 1      35.1 172,142      91.5  17,626        9.4 170,414      90.6 161,086
#>   pctOver21  Over25 pctOver25 Over62 pctOver62 Over65 pctOver65   Males
#> 1      85.7 145,709      77.5 24,920      13.3 19,899      10.6  90,316
#>   pctMales Over65Males pctOver65Males Females pctFemales Over65Females
#> 1     48.0       8,792            9.7  97,724       52.0        11,107
#>   pctOver65Females OneRace pctOneRace  White1 pctWhite1 Black1 pctBlack1
#> 1             11.4 182,709       97.2 121,324      64.5 24,103      12.8
#>   Indian1 pctIndian1  Asian1 pctAsian1 HawnPI1 pctHawnPI1 Other1 pctOther1
#> 1     407        0.2  31,688      16.9      33        0.0  5,154       2.7
#>   MultiRace pctMultiRace  White2 pctWhite2 Black2 pctBlack2 Indian2 pctIndian2
#> 1     5,331          2.8 125,685      66.8 26,033      13.8   1,128        0.6
#>    Asian2 pctAsian2 HawnPI2 pctHawnPI2 Other2 pctOther2 HispanicPop
#> 1  34,543      18.4     265        0.1  5,900       3.1      16,683
#>   pctHispanicPop NonHispPop pctNonHispPop NonHispWhite pctNonHispWhite
#> 1            8.9    171,357          91.1      111,193            59.1
#>   NonHispBlack pctNonHispBlack NonHispAmInd pctNonHispAmInd NonHispAsian
#> 1       23,655            12.6          301             0.2       31,471
#>   pctNonHispAsian NonHispHawnPI pctNonHispHawnPI  TotHHs HHInc0 pctHHInc0
#> 1            16.7            20              0.0 104,774 10,394       9.9
#>   HHInc10 pctHHInc10 HHInc15 pctHHInc15 HHInc25 pctHHInc25 HHInc35 pctHHInc35
#> 1   3,847        3.7   5,771        5.5   5,371        5.1   7,579        7.2
#>   HHInc50 pctHHInc50 HHInc75 pctHHInc75 HHInc100 pctHHInc100 HHInc150
#> 1  14,927       14.2  12,558       12.0   17,089        16.3   10,232
#>   pctHHInc150 HHInc200 pctHHInc200 NumHHEarnings pctNumHHEarnings NumHHSocSec
#> 1         9.8   17,006        16.2        89,574             85.5      14,338
#>   pctNumHHSocSec NumHHRetInc pctNumHHRetInc NumHHSuppSecInc pctNumHHSuppSecInc
#> 1           13.7       7,799            7.4           2,614                2.5
#>   NumHHPubAssist pctNumHHPubAssist MedianHHInc AvgHHInc AvgHHEarnings
#> 1          1,297               1.2     $87,213 $126,098      $125,974
#>   AvgHHSocSec AvgHHRetInc AvgHHSuppSecInc AvgHHPubAssist FamHHs pctFamHHs
#> 1     $19,032     $43,492          $8,521         $1,501 34,655      33.1
#>   FamHHInc0 pctFamHHInc0 FamHHInc10 pctFamHHInc10 FamHHInc15 pctFamHHInc15
#> 1     1,432          4.1        557           1.6      1,220           3.5
#>   FamHHInc25 pctFamHHInc25 FamHHInc35 pctFamHHInc35 FamHHInc50 pctFamHHInc50
#> 1      1,445           4.2      1,694           4.9      3,077           8.9
#>   FamHHInc75 pctFamHHInc75 FamHHInc100 pctFamHHInc100 FamHHInc150
#> 1      3,058           8.8       6,308           18.2       5,132
#>   pctFamHHInc150 FamHHInc200 pctFamHHInc200 AvgFamInc     PCI NonFamHHs
#> 1           14.8      10,732           31.0  $192,455 $71,381    70,119
#>   pctNonFamHHs MedianNonFamInc AvgNonFamInc PovUniverseFamilyPop PoorInFamily
#> 1         66.9         $68,941      $93,154               89,483        7,338
#>   pctPoorInFamily PoorFamilies pctPoorFamilies PovRatioUnderHalf
#> 1             8.2        2,330             6.7            16,571
#>   PovRatiov5tov99 PovRatio1to2 PovRatioOver2  Over16 LaborForce pctLaborForce
#> 1           9,658       15,657       137,365 171,585    128,146          7769
#>   CivLabForce pctCivLabForce EmployedCLF pctEmployedCLF UnemployedCLF
#> 1     128,030           7761     119,911          10001         8,119
#>   pctUnemployedCLF Military pctMilitary NotInLF pctNotInLF OwnKidsUnder6
#> 1            799.4      116         7.8  43,439       3031         9,442
#>   OwnKidsU6AllPrntsWk pctOwnKidsU6AllPrntsWk OwnKidsOver6 OwnKidsO6AllPrntsWk
#> 1               6,132                   64.9        7,548               5,821
#>   pctOwnKidsO6AllPrntsWk Worker16 Commuters pctCommuters DriveAlone
#> 1                   77.1  117,910   109,622         93.0     36,632
#>   pctDriveAlone Carpool pctCarpool PublicTrans pctPublicTrans WalkToWork
#> 1          31.1   4,729        4.0      28,481           24.2     33,049
#>   pctWalkToWork OtherCommute pctOtherCommute WorkAtHome pctWorkAtHome
#> 1          28.0        6,731             5.7      8,288           7.0
#>   ManProfOccs pctManProfOccs ServiceOccs pctServiceOccs SalesOffOccs
#> 1      79,365           66.2      10,772            9.0       25,565
#>   pctSalesOffOccs FarmFishOccs pctFarmFishOccs ConsOccs pctConsOccs TransOccs
#> 1            21.3            0             0.0    1,231         1.0     2,978
#>   pctTransOccs Agriculture pctAgriculture Construction pctConstruction
#> 1          2.5         195            0.2        1,653             1.4
#>   Manufacturing pctManufacturing WholesaleTrade pctWholesaleTrade RetailTrade
#> 1         6,325              5.3          2,859               2.4       6,795
#>   pctRetailTrade Transportation pctTransportation Information pctInformation
#> 1            5.7          3,670               3.1       3,641            3.0
#>   Finance_Ins pctFinance_Ins Professional pctProfessional Educ_Health_SocSvcs
#> 1      19,004           15.8       29,514            24.6              29,199
#>   pctEduc_Health_SocSvcs RecreationEtc pctRecreationEtc OtherIndustries
#> 1                   24.4         9,289              7.7           4,525
#>   pctOtherIndustries PublicAdmin pctPublicAdmin PrivWageWorkers
#> 1                3.8       3,242            2.7         104,016
#>   pctPrivWageWorkers GovWorkers pctGovWorkers SelfEmpWorkers pctSelfEmpWorkers
#> 1               86.7     10,804           9.0          5,075               4.2
#>   UnpaidFamWorkers pctUnpaidFamWorkers Families pctFamilies FamsWithKids
#> 1               16                 0.0   34,655        33.1       10,187
#>   pctFamsWithKids MarriedCouples pctMarriedCouples CouplesWithKids
#> 1             9.7         28,058              26.8           7,441
#>   pctCouplesWithKids SingleMaleFamilies pctSingleMaleFamilies SingleFathers
#> 1                7.1              1,760                   1.7           451
#>   pctSingleFathers SingleFemaleFamilies pctSingleFemaleFamilies SingleMothers
#> 1              0.4                4,837                     4.6         2,295
#>   pctSingleMothers HHsWithKids pctHHsWithKids HHsWithElders pctHHsWithElders
#> 1              2.2      10,982           10.5        15,175             14.5
#>   AvgHHSize AvgFamSize   HHPop pctHHPop FamHHpop pctFamHHpop NonFamHHpop
#> 1      1.70       2.58 178,366     94.9   90,671        5453      87,695
#>   pctNonFamHHpop GrpQuarters pctGrpQuarters Householder pctHouseholder Spouse
#> 1           5347       9,674            5.1     104,774           6396 28,130
#>   pctSpouse  Child pctChild OtherRelative pctOtherRelative NonRelative
#> 1      1694 21,372     1251         5,326            344.8      18,764
#>   pctNonRelative UnmarriedPartner pctUnmarriedPartner UMPartnerHHsPerK
#> 1           1114            7,350               442.4               70
#>   NeverMarried pctNeverMarried Married pctMarried Separated pctSeparated
#> 1       89,738            52.1  60,712       35.3     1,809          1.0
#>   Widowed pctWidowed Divorced pctDivorced   Over3 EnrolledOver3
#> 1   5,594        3.2   14,289         8.3 182,110        41,076
#>   pctEnrolledOver3 InNursery pctInNursery InKindergarten pctInKindergarten
#> 1             22.6     2,273          5.5            802               2.0
#>   InElementary pctInElementary InHighSchool pctInHighSchool InCollege
#> 1        5,888            14.3        2,047             5.0    30,066
#>   pctInCollege LessThan9th pctLessThan9th SomeHighSchool pctSomeHighSchool
#> 1         73.2       3,370            2.3          3,382               2.3
#>   HighSchool pctHighSchool SomeCollege pctSomeCollege Associates pctAssociates
#> 1      9,366           6.4      15,490           10.6      4,914           3.4
#>   Bachelors pctBachelors GradProf pctGradProf HighSchoolOrMore
#> 1    56,679         38.9   52,508        36.0          138,957
#>   pctHighSchoolOrMore Bachelorsormore pctBachelorsormore Civilian pctCivilian
#> 1                95.4         109,187               74.9  170,298       148.1
#>   Veteran pctVeteran  TotHUs  OccHUs pctOccHUs OwnerOcc pctOwnerOcc RenterOcc
#> 1   4,984        2.9 121,960 104,774      85.9   43,115        41.2    61,659
#>   pctRenterOcc AvgOwnerHHSize AvgRenterHHSize VacHUs pctVacHUs VacantForSale
#> 1         58.8           1.77            1.65 17,186      14.1         2,010
#>   pctVacantForSale VacantForRent pctVacantForRent VacantSeasonal
#> 1             11.7         4,068             23.7          6,416
#>   pctVacantSeasonal TotalOwnerUnits OwnerVacRate TotalRentalUnits RenterVacRate
#> 1              37.3          45,125          4.5           65,727           6.2
#>   PersonsInOwnerUnits pctPersonsInOwnerUnits PersonsInRenterUnits
#> 1              76,424                   40.6              101,942
#>   pctPersonsInRenterUnits Units1 pctUnits1 Units1Detached pctUnits1Detached
#> 1                    54.2  8,900       7.3          3,064              34.4
#>   Units1Attached pctUnits1Attached Units2 pctUnits2 Units3_4 pctUnits3_4
#> 1          5,836              65.6  2,314       1.9    4,395         3.6
#>   Units5_9 pctUnits5_9 Units10_19 pctUnits10_19 Units20up pctUnits20up
#> 1    4,719         3.9      2,906           2.4    98,556         80.8
#>   MobileHomes pctMobileHomes BoatRV pctBoatRV MobileHomesPerK Built2010orLater
#> 1         161            0.1      9       0.0               1            3,562
#>   pctBuilt2010orLater Built2000_2009 pctBuilt2000_2009 Built1990_1999
#> 1                 2.9         40,553              33.2         13,964
#>   pctBuilt1990_1999 Built1980_1989 pctBuilt1980_1989 Built1970_1979
#> 1              11.4         13,116              10.7         11,357
#>   pctBuilt1970_1979 Built1960_1969 pctBuilt1960_1969 Built1950_1959
#> 1               9.3          9,430               7.7          5,083
#>   pctBuilt1950_1959 Built1940_1949 pctBuilt1940_1949 BuiltBefore1940
#> 1               4.2          3,022               2.5          21,873
#>   pctBuiltBefore1940 MovedIn2010orLater pctMovedIn2010orLater MovedIn2000_2009
#> 1               17.9             55,816                  53.3           37,134
#>   pctMovedIn2000_2009 MovedIn1990_1999 pctMovedIn1990_1999 MovedIn1980_1989
#> 1                35.4            7,343                 7.0            2,556
#>   pctMovedIn1980_1989 MovedInBefore1980 pctMovedInBefore1980 NoVehicles
#> 1                 2.4             1,925                  1.8     37,797
#>   pctNoVehicles Vehicles1 pctVehicles1 Vehicles2 pctVehicles2 VehiclesGE3
#> 1          36.1    53,792         51.3    11,706         11.2       1,479
#>   pctVehiclesGE3 HHFUtilGas pctHHFUtilGas HHFLPGas pctHHFLPGas HHFElectric
#> 1            1.4     57,711          55.1      790         0.8      43,786
#>   pctHHFElectric HHFKerosene pctHHFKerosene HHFCoal pctHHFCoal HHFWood
#> 1           41.8         201            0.2      44        0.0      48
#>   pctHHFWood HHFSolar pctHHFSolar HHFOther pctHHFOther HHFNoFuel pctHHFNoFuel
#> 1        0.0       17         0.0      791         0.8     1,386          1.3
#>   NoPlumbing pctNoPlumbing NoKitchen pctNoKitchen NoPhone pctNoPhone
#> 1        546           0.5       757          0.7   2,992        2.9
#>   PersonsPerRoomLow pctPersonsPerRoomLow PersonsPerRoomMedium
#> 1           102,591                 97.9                1,030
#>   pctPersonsPerRoomMedium PersonsPerRoomHigh pctPersonsPerRoomHigh HvalUnder50
#> 1                     1.0              1,153                   1.1         514
#>   pctHvalUnder50 Hval50 pctHval50 Hval100 pctHval100 Hval150 pctHval150 Hval200
#> 1            1.2    564       1.3     970        2.2   2,430        5.6   9,825
#>   pctHval200 Hval300 pctHval300 Hval500 pctHval500 HvalOverMillion
#> 1       22.8  15,754       36.5   8,691       20.2           2,907
#>   pctHvalOverMillion HvalOver2Million pctHvalOver2Million MedianHValue
#> 1                6.7            1,460                 3.4     $430,285
#>   AvgHValue HUsMort pctHUsMort HUsMortOver30Pct pctHUsMortOver30Pct
#> 1  $546,029  31,880       74.0           10,689                33.5
#>   MedianOwnerCostsMort HUsNoMort pctHUsNoMort HUsNoMortOver30Pct
#> 1               $2,537    11,235         26.0              2,864
#>   pctHUsNoMortOver30Pct MedianOwnerCostsNoMort CashRenter pctCashRenter
#> 1                  25.5                 $1,182     59,929          97.2
#>   NoCashRenter pctNoCashRenter MedianGrossRent AvgGrossRent CashRenterOver30Pct
#> 1        1,730             2.8          $1,601       $1,635              25,807
#>   pctCashRenterOver30Pct CashRenterOver750 pctCashRenterOver750 Persons_Sampled
#> 1                   41.9            44,821                 72.7           10727
#>   PctPersons_Sampled HUs_Sampled PctHUs_Sampled PctOver5 PctOver18Males
#> 1                5.7        6044            5.0     95.4           89.9
#>   PctOver18Females PovUniverse  poor Pctpoor PctPovRatioUnderHalf
#> 1             91.3        1396 26229    14.6                  9.2
#>   PctPovRatiov5tov99 PctPovRatio1to2 PctpovratioOver2
#> 1                5.4             8.7             76.6
#>  [ reached getOption("max.print") -- omitted 1 row ]